Move uncertainty to ml.fit and cleanup code
